implement logging decorator for improved script tracing
