+++ tsconfig.json
@@ -8,6 +8,7 @@
     "sourceMap": true,
     "esModuleInterop": true,
     "resolveJsonModule": true,
+    "experimentalDecorators": true,
     "allowSyntheticDefaultImports": true,
     "moduleResolution": "node"
   },
